Change sublime-menu, plugin logger, commands, search_syntaxes and icon_syntaxes.

+++ src/zukan_icon_theme/commands/commands.py
@@ -1,5 +1,6 @@
 import logging
 import os
+import sublime
 import sublime_plugin
 
 from ..lib.icons_syntaxes import ZukanSyntax
@@ -8,6 +9,7 @@ from ..lib.themes import ThemeFile
 from ..helpers.search_themes import search_resources_sublime_themes
 from ..utils.zukan_dir_paths import (
     ZUKAN_INSTALLED_PKG_PATH,
+    ZUKAN_PKG_ICONS_PATH,
 )
 
 logger = logging.getLogger(__name__)
@@ -20,7 +22,11 @@ class DeleteTheme(sublime_plugin.TextCommand):
 
     def run(self, edit, theme_name: str):
         # print(theme_name)
-        ThemeFile.delete_created_theme_file(theme_name)
+        message = 'Are you sure you want to delete {t}?'.format(
+            t=os.path.join(ZUKAN_PKG_ICONS_PATH, theme_name)
+        )
+        if sublime.yes_no_cancel_dialog(message) == sublime.DIALOG_YES:
+            ThemeFile.delete_created_theme_file(theme_name)
 
     def input(self, args: dict):
         # print(args)
@@ -42,7 +48,9 @@ class DeleteThemeInputHandler(sublime_plugin.ListInputHandler):
         if ThemeFile.list_created_themes_files():
             return sorted(ThemeFile.list_created_themes_files())
         else:
-            logger.info('it does not exist any created theme, list is empty')
+            raise TypeError(
+                logger.info('it does not exist any created theme, list is empty')
+            )
 
 
 class DeleteThemes(sublime_plugin.ApplicationCommand):
@@ -52,9 +60,15 @@ class DeleteThemes(sublime_plugin.ApplicationCommand):
 
     def run(self):
         if ThemeFile.list_created_themes_files():
-            ThemeFile.delete_created_themes_files()
+            message = 'Are you sure you want to delete all themes in {f}?'.format(
+                f=ZUKAN_PKG_ICONS_PATH
+            )
+            if sublime.yes_no_cancel_dialog(message) == sublime.DIALOG_YES:
+                ThemeFile.delete_created_themes_files()
         else:
-            logger.info('it does not exist any created theme, list is empty')
+            raise TypeError(
+                logger.info('it does not exist any created theme, list is empty')
+            )
 
 
 class InstallTheme(sublime_plugin.TextCommand):
@@ -62,10 +76,10 @@ class InstallTheme(sublime_plugin.TextCommand):
     Sublime command to create theme from a list of installed themes.
     """
 
-    def run(self, edit, theme_name):
+    def run(self, edit, theme_name: str):
         ThemeFile.create_theme_file(theme_name)
 
-    def input(self, args):
+    def input(self, args: dict):
         return InstallThemeInputHandler()
 
 
@@ -90,7 +104,9 @@ class InstallThemeInputHandler(sublime_plugin.ListInputHandler):
         if list_themes_not_installed:
             return sorted(list_themes_not_installed)
         else:
-            logger.info('all themes are already created, list is empty.')
+            raise TypeError(
+                logger.info('all themes are already created, list is empty.')
+            )
 
 
 class InstallThemes(sublime_plugin.ApplicationCommand):
@@ -114,7 +130,10 @@ class RebuildFiles(sublime_plugin.ApplicationCommand):
         Try move folders if installed trough Package Control. Then install
         sublime-theme and sublime-syntax files.
         """
-        if os.path.exists(ZUKAN_INSTALLED_PKG_PATH):
-            MoveFolder.move_folders()
-        ZukanSyntax.create_icons_syntaxes()
-        ThemeFile.create_themes_files()
+        try:
+            ZukanSyntax.remove_icons_syntaxes()
+            if os.path.exists(ZUKAN_INSTALLED_PKG_PATH):
+                MoveFolder.move_folders()
+        finally:
+            ZukanSyntax.create_icons_syntaxes()
+            ThemeFile.create_themes_files()
